+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Changes prior to the next official version change will appear here.
     * bugfix in find_symbol tool (a bug fixed in LS)
     * merged the two overview tools (for dir and file) int a single one
     * one-click setup for Cline enabled
+    * search for pattern tool can now (optionally) search in the entire project
+    * new tool for restarting the language server, in case of other sources of editing apart from Serena
     * Fix `CheckOnboardingPerformedTool`:
         * Tool description was incompatible with project change
         * Returned result was not as useful as it could be (now added list of memories)